fix: resolve SwiftLint violations in Plex auth code
Extract Plex auth UI into separate PlexAuthView and MobilePlexAuthView
to fix type_body_length violations. Fix force unwraps and optional
boolean warnings in PlexClient.
